Added API config
Added API config behaviour. The system can now be configured by sending json documents via text TCP. These settings are going to be used to create presets and instances of modules and setups for the UI panels connected to the system. The settings are separated by type according to the user input.